Reorganiza estructura de archivos de carga de imagen
