+++ vin/modelo.py
@@ -1,34 +1,6 @@
 from vin.vin_image import VinImage
 from vin.utils import distancia
-import os
 from typing import List
-
-def cargar_imagenes(directorio: str) -> List[VinImage]:
-    """Carga la estructura de imágenes, almacenada de la siguiente forma:
-    ```
-    img
-    |-> clase1
-    |      |-> archivo1.png
-    |      |
-    |      ...
-    |-> clase2
-    ...
-    ```
-    """
-    imagenes = []
-
-    print("Comienza la carga de imágenes...")
-    for subcarpeta in os.scandir(directorio):
-        print(f'\t→ Cargando categoría {subcarpeta.name}')
-        for img in os.scandir(subcarpeta):
-            if (img.is_file):
-                imagenes.append(
-                    VinImage(img.path, subcarpeta.name)
-                )
-
-    print("Listo ✓")
-    return imagenes
-
 
 def knn(imagen: VinImage, dataset: List[VinImage], k: int) -> str:
     """Aproxima `imagen` a los `k` vecinos más próximos utilizando las imágenes presentes en `dataset`
